requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==4.2.16
django-bootstrap-v5==1.0.11
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
soupsieve==2.6
sqlparse==0.5.1
tzdata==2024.2
